creating rollback

+++ src/main/java/kubernetes/client/controller/ProjectController.java
@@ -5,9 +5,6 @@ import java.util.List;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -17,10 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
-import kubernetes.client.model.Customer;
 import kubernetes.client.model.Project;
-import kubernetes.client.security.AuthenticatedUser;
-import kubernetes.client.service.CustomerService;
 import kubernetes.client.service.ProjectService;
 import kubernetes.client.validator.ProjectValidator;
 
@@ -29,9 +23,6 @@ public class ProjectController extends BaseController {
 
 	@Autowired
 	private ProjectService projectService;
-
-	@Autowired
-	private CustomerService customerService;
 
 	@Autowired
 	private ProjectValidator projectValidator;
@@ -61,7 +52,6 @@ public class ProjectController extends BaseController {
 		}
 
 		projectService.insert(project, getCurrentUser().getCustomer().getId());
-		// setCurrentUser();
 		redirectAttributes.addFlashAttribute("info", "Project created successfully");
 		return "redirect:/projects";
 	}
@@ -83,20 +73,16 @@ public class ProjectController extends BaseController {
 		return "redirect:/projects";
 	}
 
-	@RequestMapping(value = "/projects/{id}", method = RequestMethod.DELETE)
+	@RequestMapping(value = "/projects/delete/{id}", method = RequestMethod.GET)
 	public String deleteProject(@PathVariable int id, Model model, RedirectAttributes redirectAttributes) {
+		Project project = projectService.getProjectById(id, getCurrentUser().getCustomer().getId());
+		if (project == null) {
+			model.addAttribute("error", "The Project does not exist or you are not authorized to delete it.");
+			return "403";
+		}
 		projectService.delete(id);
-		//setCurrentUser();
 		redirectAttributes.addFlashAttribute("info", "Project deleted successfully");
 		return "redirect:/projects";
-	}
-
-	public void setCurrentUser() {
-		Customer customer = customerService.getCustomerByEmail(getCurrentUser().getCustomer().getEmail());
-		AuthenticatedUser user = new AuthenticatedUser(customer);
-		Authentication authentication = new UsernamePasswordAuthenticationToken(user.getCustomer().getEmail(),
-				user.getCustomer().getPassword(), user.getAuthorities());
-		SecurityContextHolder.getContext().setAuthentication(authentication);
 	}
 
 }
